requirements.txt
Faker==25.2.0
oracledb==2.2.0
python-dotenv==1.0.1
APScheduler==3.10.4
colorama==0.4.6
python-dotenv